requirements.txt
selenium
loguru
aiohttp
asyncio
click
beautifulsoup4
python-dotenv
faker